The PlotResults function was changed to be a method of the CmlJobManager class, since other classes might have similar plotting functionality.
The PlotResults method of the CmlJobManager class was modified to add the time of the printing of the figure as its title.

The generic JobManager class was modified to work more efficiently and to generate less messages.

The generic JobManager class was modified so that it does not generate any new tasks while the first round of short tasks for a new job are not yest finished. This is to prevent over-simulation in low SNR points for CML project.