The generic JobManager was modified to be able to handle the requirements of the new PlbpJobManager. In particular, a feature was added to support the case in which the job could be fully finished by the job manager itself without requiring to generate any tasks. A flag called 'JobStatus' is set in the JobState structure by the PreProcess method. The JobManager will move the input job finished by the Preprocess method from the user's JobIn directory to his JobOut directory upon observing this flag.
Moreover, a feature was added that the job manager could update the job file after generating a (set of) tasks (and before any tasks are returned in the user's TaskOut directory). This could be achieved by setting a flag called 'JobFileUpdateRequest' in the JobState structure by the 'DivideJob2Tasks' method. This feature would be useful if only one set of tasks are generated for a job and the number of required tasks is deterministic and could be known before running any tasks. The 'DivideJob2Tasks' method can set a field in the JobState that signals the required tasks have been generated and no more tasks are needed. Since the updated job is saved by the JobManager, the next time that it is loaded, this method knows that no more tasks are required to be generated.